requirements.txt
# FastAPI y servidor
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7

# AWS y almacenamiento
boto3==1.40
botocore==1.40

# Procesamiento de datos
pandas==2.3.2
numpy==2.2.6

# Utilidades
python-dotenv==1.1.1
python-multipart==0.0.6

# LlamaIndex
llama-index==0.10.11
llama-index-llms-groq==0.1.3